chore: refactoring enums classes with enum trait
