+++ app/Enums/UgcValidatedStatus.php
@@ -2,30 +2,13 @@
 
 namespace App\Enums;
 
-use ReflectionClass;
+use App\Traits\EnumsTrait;
 
 class UgcValidatedStatus
 {
+    use EnumsTrait;
+
     const Valid = 'valid';
     const Invalid = 'invalid';
     const NotValidated = 'not_validated';
-
-
-    public static function cases()
-    {
-        $self = new self();
-        $constants = $self->getConstants();
-
-        $cases = [];
-        foreach ($constants as $constant) {
-            $cases[$constant] = $constant;
-        }
-
-        return $cases;
-    }
-    public function getConstants()
-    {
-        $reflectionClass = new ReflectionClass($this);
-        return $reflectionClass->getConstants();
-    }
 }
